Shorten build script by removing unecessary indentation.

+++ _build.py
@@ -8,8 +8,7 @@ scriptsDir = "scripts"
 imagesDir = "img"
 
 # Create output site directory
-if os.path.isdir(outDir):
-    shutil.rmtree(outDir)
+if os.path.isdir(outDir): shutil.rmtree(outDir)
 os.makedirs(outDir)
 shutil.copytree(imagesDir, f"{outDir}/{imagesDir}")
 shutil.copytree(stylesDir, f"{outDir}/{stylesDir}")
@@ -17,11 +16,4 @@ shutil.copytree(scriptsDir, f"{outDir}/{scriptsDir}")
 shutil.copyfile("google9c66b3b3d14f628e.html", f"{outDir}/google9c66b3b3d14f628e.html")
 
 # Create index page
-subprocess.run([
-    "pandoc",
-    "--from=markdown",
-    "--to=html",
-    f"--output={outDir}/index.html",
-    "--template=templates/index.html",
-    "index.md"
-])
+subprocess.run(["pandoc", "--from=markdown", "--to=html", f"--output={outDir}/index.html", "--template=templates/index.html", "index.md"])
